ui tweaks

+++ LM-Kit-Maestro/UI/UIConstants.cs
@@ -4,7 +4,7 @@ namespace LMKit.Maestro.UI;
 
 public static class UIConstants
 {
-    public const double ChatWindowLayoutMinimumWidth = 720;
+    public const int ChatWindowLayoutMinimumWidth = 720;
 
     public const double WindowMinimumWidth = 568;
